+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- State's pendingTransactions does not reference firebase (for better backend interoperability)
+
 ### Added
 
 - Added a change log